On Windows, use win32api.GetUserName instead of getpass.getuser
getpass.getuser() should not be used for authentication because a
malicious user may simply change an environment variable (eg LOGNAME)
and impersonate any user. Due to this, use win32api.GetUserName()
which is a more secure way of obtaining the user name on Windows.

'pwd' was imported twice. Removed one instance. Use 'HAS_PWD' as
the presence check.

Signed-off-by: Sergey Kizunov <sergey.kizunov@ni.com>